move scripts to Runtime
