--- UniRx.meta
@@ -1,3 +1,0 @@
-﻿fileFormatVersion: 2
-guid: 565b512b58d544138f53da563c097f38
-timeCreated: 1566734723
