--- UniRx/AxisInputUtil.cs.meta
@@ -1,3 +1,0 @@
-﻿fileFormatVersion: 2
-guid: 7041168c83fd4461af1a386bb482f49d
-timeCreated: 1566740438
