--- UniRx/KeyInputUtil.cs.meta
@@ -1,3 +1,0 @@
-﻿fileFormatVersion: 2
-guid: fa1e54a6c66942d9a6409e7e997e02a0
-timeCreated: 1566740612
