--- UniRx/MouseButtonInputUtil.cs.meta
@@ -1,3 +1,0 @@
-﻿fileFormatVersion: 2
-guid: 8d3cd9d7740a4c67a4861d177d1f7cfa
-timeCreated: 1566740641
